% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_data}
\alias{combined_data}
\title{Cross-situational word learning experiment data}
\format{
\subsection{\code{combined_data}}{

A list of 44 experimental conditions with trial orders and human performance:
\describe{
\item{condition}{3x4}
\item{condition}{3x4 1/.5}
\item{condition}{3x4 1/.66}
\item{condition}{3x4 +6o}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
combined_data
}
\description{
A collection of data from human cross-situational word learning experiments,
including the trial-by-trial design of each experiment and the item-level
proportion of correct responses from a sample of N participants.
}
\keyword{datasets}
