% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_cooc_matrix}
\alias{create_cooc_matrix}
\title{Creates Co-occurrence Matrix from Training Trials}
\usage{
create_cooc_matrix(train)
}
\arguments{
\item{train}{A list representing the training data, where each element is a
trial that includes both words and objects. The structure is expected to
have sub-elements \code{words} and \code{objs} for each trial.}
}
\value{
A matrix where each element represents the count of co-occurrences
between a word (rows) and an object (columns). The row and column names
correspond to the unique words and objects found in the training data,
respectively.
}
\description{
Given a training order (list of words and objects per trial), this function
returns a matrix tallying word-object co-occurrences across the trials. This
matrix can be used to analyze the frequency with which certain words and
objects appear together during the training phase.
}
