% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-decay.R
\name{decay}
\alias{decay}
\title{Decay model}
\usage{
decay(C)
}
\arguments{
\item{C}{decay}
}
\value{
An object of class xslMod
}
\description{
Decay model
}
\examples{
mod <- decay(C = 0.5)
xsl_run(mod, get_example_ambiguous_condition())
}
