% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fazly.R
\name{fazly}
\alias{fazly}
\title{Fazly et al. 2010 probablistic associative model}
\usage{
fazly(lambda, beta, alpha = 10, epsilon = 0.001)
}
\arguments{
\item{lambda}{Small smoothing factor (e.g. 1e-5)}

\item{beta}{upper bound on number of symbol types to expect (e.g. 8500)}

\item{alpha}{fixed smoothing parameter (e.g. 10)}

\item{epsilon}{fixed smoothing parameter (e.g. 0.001)}
}
\value{
An object of class xslMod
}
\description{
Fazly et al. 2010 probablistic associative model
}
