% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_example_ambiguous_condition}
\alias{get_example_ambiguous_condition}
\title{Get example ambiguous condition}
\usage{
get_example_ambiguous_condition()
}
\value{
A list of trials (with nested words and objects per trial) and perf (P(correct referent | word))
}
\description{
This function creates an example condition with two trials, and two word-object pairs
per trial. Each word-object pair appears only once, and thus each word has a 50\% chance of
being associated with the correct referent. The function returns a list of the trials
with nested words and objs per trial, as well as a vector of the conditional probability
of correctly selecting the correct referent given each word.
}
