% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_model_dataframe}
\alias{get_model_dataframe}
\title{Create a Data Frame from Model Fits}
\usage{
get_model_dataframe(fits, conds, cvdat = c())
}
\arguments{
\item{fits}{A list of model fitting results, where each element corresponds to a different model. The elements are objects resulting from the optimization process (e.g., using \code{DEoptim}).}

\item{conds}{A list representing the conditions for which the models were fitted. Each element in the list should correspond to a condition and include relevant data such as human item accuracy (\code{HumanItemAcc}).}

\item{cvdat}{An optional list of cross-validation data. If provided, this data is used instead of the data in \code{conds}. Default is an empty vector, indicating that \code{conds} will be used.}
}
\value{
A \code{tibble} (data frame) where each row represents a model-condition pair and includes columns for the model name, condition number, condition details, model performance, human performance, and subject count.
}
\description{
This function generates a data frame summarizing the fits of models to various conditions. It compiles information such as model performance, human performance, and subject count for each condition, creating a comprehensive overview suitable for analysis and visualization.
}
