% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_model_dataframe_cond_fits}
\alias{get_model_dataframe_cond_fits}
\title{Generate Data Frame for Model Fits by Condition}
\usage{
get_model_dataframe_cond_fits(fits, conds)
}
\arguments{
\item{fits}{A list of model fitting results by condition, where each top-level element corresponds to a different model and each nested element represents a specific condition.}

\item{conds}{A list representing the conditions for which the models were fitted, including data such as human item accuracy (\code{HumanItemAcc}).}
}
\value{
A \code{tibble} (data frame) where each row represents a specific model-condition pair, including columns for the model name, condition number, detailed condition information, model performance, human performance, and number of subjects.
}
\description{
This function creates a data frame summarizing the fits of models for each specific condition. It consolidates data such as model performance, human performance, and number of subjects for each condition, enabling detailed analysis and comparison of model fits across different conditions.
}
