% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_roc_max}
\alias{get_roc_max}
\title{Get Maximum F-score from ROC Scores}
\usage{
get_roc_max(mdat, gold_lexicon = c())
}
\arguments{
\item{mdat}{A matrix representing the model association matrix with words as
rows and referents as columns, used to calculate the ROC scores.}

\item{gold_lexicon}{Optional; a data frame or list where each row/element
represents a word-object pair in the gold lexicon. If provided, it enhances
the accuracy of the ROC score calculations.}
}
\value{
A single numeric value representing the maximum F-score obtained from
the ROC scores of the model.
}
\description{
This function computes the maximum F-score from the Receiver Operating
Characteristic (ROC) scores of a model. It leverages the \code{get_roc} function
to calculate the ROC scores and then extracts the highest F-score, providing
a concise metric for the best classification performance of the model.
}
