% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_tp}
\alias{get_tp}
\title{Get True Positives (TP), Given a Knowledge Matrix and a Gold Lexicon}
\usage{
get_tp(m, gold_lexicon)
}
\arguments{
\item{m}{A matrix representing the knowledge matrix with words as rows and
referents as columns.}

\item{gold_lexicon}{A data frame or list where each row/element represents a
word-object pair in the gold lexicon.}
}
\value{
A single value with the expected number of true positives.
}
\description{
This function iterates over words in a given gold lexicon and accumulates the
associative strength (can be integral e.g. 1, or real-valued) in the
knowledge matrix for the intended referents (present in the gold lexicon).
Returns the number of expected true positives (TP) for this gold lexicon and
knowledge matrix.
}
