% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\title{Cross-situational word learning models}
\format{
\subsection{\code{models}}{

A list of
\describe{
\item{model}{baseline}
\item{model}{Bayesian_decay}
\item{model}{decay}
\item{model}{fazly}
\item{model}{fazlyt}
\item{model}{kachergis}
\item{model}{novelty}
\item{model}{rescorla-wagner}
\item{model}{guess-and-test}
\item{model}{kachergis_sampling}
\item{model}{multi_sampling}
\item{model}{propose-but-verify}
\item{model}{pursuit}
\item{model}{strength}
\item{model}{tilles}
\item{model}{uncertainty}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
models
}
\description{
A collection of cross-situational word learning models that learn
trial-by-trial.
}
\keyword{datasets}
