% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-multi_sampling.R
\name{multi_sampling}
\alias{multi_sampling}
\title{Multi-sampling associative model}
\usage{
multi_sampling(C, X, B, K = 1)
}
\arguments{
\item{C}{Decay}

\item{X}{Total associative weight to distribute per trial}

\item{B}{Associative weight to distribute each sample}

\item{K}{Number of associations to update per word (constraint: B <= X)}
}
\value{
An object of class xslMod
}
\description{
Multi-sampling associative model
}
\examples{
mod <- multi_sampling(C = 1, X = .2, B = .3, K = 1)
xsl_run(mod, get_example_ambiguous_condition())
}
