% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-propose_but_verify.R
\name{propose_but_verify}
\alias{propose_but_verify}
\title{Trueswell et al. 2013 propose-but-verify model}
\usage{
propose_but_verify(alpha, alpha_increase)
}
\arguments{
\item{alpha}{initial association strength}

\item{alpha_increase}{learning rate}
}
\value{
An object of class xslMod
}
\description{
Trueswell et al. 2013 propose-but-verify model
}
\examples{
mod <- propose_but_verify(alpha = .1, alpha_increase = .5)
xsl_run(mod, get_example_ambiguous_condition())
}
