% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-pursuit.R
\name{pursuit}
\alias{pursuit}
\title{Stevens et al. 2014 pursuit model}
\usage{
pursuit(gamma, threshold, lambda)
}
\arguments{
\item{gamma}{learning rate}

\item{threshold}{threshold (probability) to move an association to the known lexicon}

\item{lambda}{smoothing probability}
}
\value{
An object of class xslMod
}
\description{
Stevens et al. 2014 pursuit model
}
\examples{
mod <- pursuit(gamma = .2, threshold = .3, lambda = .05)
xsl_run(mod, get_example_ambiguous_condition())
}
