% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rescorla_wagner.R
\name{rescorla_wagner}
\alias{rescorla_wagner}
\title{Rescorla-Wagner (1972) error-driven associative model}
\usage{
rescorla_wagner(C, alpha, beta, lambda)
}
\arguments{
\item{C}{Decay}

\item{alpha}{Salience (fix at 1 unless manipulated)}

\item{beta}{Learning rate -- a proportion of lambda}

\item{lambda}{Maximum associative value that a CS can achieve - should be
larger than learning rate}
}
\value{
An object of class xslMod
}
\description{
Rescorla-Wagner (1972) error-driven associative model
}
\examples{
mod <- rescorla_wagner(C = 1, alpha = 1, beta = 0.1, lambda = 3)
xsl_run(mod, get_example_ambiguous_condition())
}
