% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_model}
\alias{run_model}
\title{Run specified model on provided condition(s)}
\usage{
run_model(conds, model_name, parameters, SSE_only = F, print_perf = F)
}
\arguments{
\item{conds}{A list representing the conditions under which the model is run. Each condition should include training data (\code{train}) and human item accuracy (\code{HumanItemAcc}).}

\item{model_name}{A string specifying the name of the model to run.}

\item{parameters}{A list of parameters to be used by the model.}

\item{SSE_only}{Logical; if TRUE, only the SSE will be returned.}

\item{print_perf}{Logical; if TRUE, model performance and SSE will be printed.}
}
\value{
A list containing the model's association matrix after training, conditional probability of selecting the correct referent for each word, and the sum of squared error (SSE). If \code{SSE_only} is TRUE, only the SSE is returned.
}
\description{
This function runs a specified word learning model on a set of provided conditions. It calculates the model's performance, the sum of squared errors (SSE), and optionally, the conditional probability of selecting the correct referent for each word.
}
