% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{shannon.entropy}
\alias{shannon.entropy}
\title{Calculates Shannon entropy of a supplied vector, after normalizing it}
\usage{
shannon.entropy(p)
}
\description{
Calculates Shannon entropy of a supplied vector, after normalizing it
}
