% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{show_datasets}
\alias{show_datasets}
\title{Show available datasets in the package}
\usage{
show_datasets()
}
\value{
A data frame with dataset information
}
\description{
Returns information about the available cross-situational word learning
datasets included in the package.
}
\examples{
show_datasets()
}
