% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{show_models}
\alias{show_models}
\title{Show available models in the package}
\usage{
show_models()
}
\value{
A character vector of model names
}
\description{
Returns a list of all available cross-situational word learning models
that can be used with the package.
}
\examples{
show_models()
}
