% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{stochastic_dummy}
\alias{stochastic_dummy}
\title{Dummy Function for Stochastic Simulation}
\usage{
stochastic_dummy(n, parameters, ord)
}
\arguments{
\item{n}{Integer, representing the number of simulations to run.}

\item{parameters}{A list of parameters to be used by the model.}

\item{ord}{The order in which the model operates.}
}
\value{
A numeric value representing the model's performance.
}
\description{
This function serves as a placeholder for running stochastic simulations. It calls the \code{model} function with provided parameters and a specific order, returning the model's performance.
}
