% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tilles.R
\name{tilles}
\alias{tilles}
\title{Tilles and Fontanari (2013) reinforcement model}
\usage{
tilles(x, b, alpha_0)
}
\arguments{
\item{x}{Reinforcement parameter for stimuli in current context}

\item{b}{Inference parameter, regulating ME and prior info integration
(applies either to known words that do not appear in the current context or
to new words in the current context)}

\item{alpha_0}{Baseline efficiency corresponding to maximal uncertainty about
referent of target word}
}
\value{
An object of class xslMod
}
\description{
Tilles and Fontanari (2013) reinforcement model
}
\examples{
mod <- tilles(x = 0.6, b = 0.8, alpha_0 = 0.85)
xsl_run(mod, get_example_ambiguous_condition())
}
