% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-uncfam.R
\name{uncfam}
\alias{uncfam}
\title{Kachergis 2012}
\usage{
uncfam(X, B, C, variant = c("entropy", "novelty", "uncertainty-only"))
}
\arguments{
\item{X}{Associative weight to distribute}

\item{B}{Weighting of uncertainty vs. familiarity}

\item{C}{Decay}

\item{variant}{Which variant of the model to fit (one of "entropy",
"novelty", "uncertainty-only").}
}
\value{
An object of class xslMod
}
\description{
Kachergis et al. 2012 uncertainty- and familiarity-biased associative model
}
\examples{
mod <- uncfam(X = .1, C = 1, B = .98)
xsl_run(mod, get_example_ambiguous_condition())

mod <- uncfam(X = .1, C = 1, B = .98, variant = "novelty")
xsl_run(mod, get_example_ambiguous_condition())

mod <- uncfam(X = .1, C = 1, B = 0) # familiarity-only
xsl_run(mod, get_example_ambiguous_condition())

mod <- uncfam(X = .1, C = 1, B = .98, variant = "uncertainty-only")
xsl_run(mod, get_example_ambiguous_condition())
}
