% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-uncfam_attention.R
\name{uncfam_attention}
\alias{uncfam_attention}
\title{Biased associative model with attention scaled to trial uncertainty}
\usage{
uncfam_attention(X, B, C)
}
\arguments{
\item{X}{Associative weight to distribute}

\item{B}{Weighting of uncertainty vs. familiarity}

\item{C}{Decay}
}
\value{
An object of class xslMod
}
\description{
A variant of \code{\link[=uncfam]{uncfam()}} (Kachergis et al. 2012's uncertainty- and
familiarity-biased associative model) in which the associative weight
distributed on a trial is additionally scaled by the ratio of that
trial's mean object entropy to the mean entropy of all objects seen so
far. This implements the "system-level" theory proposed by Fitneva &
Christiansen (2015): that learners are more alert (i.e. allocate a higher
learning rate) on trials with more uncertain items, such as in a Low
Initial Accuracy condition where many items have been mis-paired.
}
\examples{
mod <- uncfam_attention(X = .1, C = 1, B = .98)
xsl_run(mod, get_example_ambiguous_condition())
}
