% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-unfam_sampling.R
\name{uncfam_sampling}
\alias{uncfam_sampling}
\title{Kachergis 2012}
\usage{
uncfam_sampling(X, B, C, K = 1)
}
\arguments{
\item{X}{Associative weight to distribute}

\item{B}{Weighting of uncertainty vs. familiarity}

\item{C}{Decay}

\item{K}{Number of associations to update per word (constraint: B <= X)}
}
\value{
An object of class xslMod
}
\description{
Kachergis et al. 2012 uncertainty- and familiarity-biased associative model
}
\examples{
mod <- uncfam_sampling(X = .1, C = 1, B = .98, K = 1)
xsl_run(mod, get_example_ambiguous_condition())
}
