% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{xsl_control}
\alias{xsl_control}
\title{Control arguments for \code{xsl_run()}}
\usage{
xsl_control(reps = 1, start_matrix = c(), test_noise = 0, n_sim = 500)
}
\arguments{
\item{reps}{Number of times to repeat training}

\item{start_matrix}{Starting matrix}

\item{test_noise}{Test noise}

\item{n_sim}{Number of simulations for stochastic models}
}
\value{
List of arguments
}
\description{
Control arguments for \code{xsl_run()}
}
