% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xsl_datasets}
\alias{xsl_datasets}
\title{Cross-situational word learning experiment data}
\format{
An object of class \code{list} of length 56.
}
\usage{
xsl_datasets
}
\description{
Cross-situational word learning experiment data
}
\keyword{datasets}
