% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslData-class.R
\name{xslData-class}
\alias{xslData-class}
\alias{xslData}
\alias{new_xslData}
\title{xslData S3 class}
\usage{
xslData(
  train = list(),
  test = list(),
  accuracy = numeric(),
  n_subj = numeric(),
  label = "",
  condition = "",
  description = "",
  response_matrix = NULL
)

new_xslData(x = list())
}
\arguments{
\item{train}{Training data}

\item{test}{Test data}

\item{accuracy}{Accuracy data}

\item{n_subj}{Number of subjects}

\item{label}{Label}

\item{condition}{Condition}

\item{description}{Description}

\item{response_matrix}{Response matrix}

\item{x}{List with elements train, test, accuracy, n_subj, label, condition, description, response_matrix}
}
\value{
An object of class xslData
}
\description{
xslData S3 class

Constructor for xslData S3 class
}
