% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslFit-class.R, R/xslMod-class.R
\name{new_xslFit}
\alias{new_xslFit}
\alias{xslMod-class}
\alias{xslMod}
\alias{new_xslMod}
\title{Constructor for xsl_model S3 class}
\usage{
new_xslFit(x = list())

xslMod(
  name = character(),
  description = character(),
  model,
  params = numeric(),
  stochastic = logical(),
  supports_start_matrix = FALSE
)

new_xslMod(x = list())
}
\arguments{
\item{x}{List with elements perf, matrix, traj, sse}

\item{name}{Name}

\item{description}{Description}

\item{model}{Model fitting function}

\item{params}{List of parameters}

\item{stochastic}{Logical indicating whether model is stochastic}

\item{supports_start_matrix}{Logical: does \code{model} initialize from
\code{control$start_matrix}? \code{xsl_run()} errors if a start matrix is supplied
to a model that doesn't, rather than letting it be silently ignored.}
}
\value{
An object of class xslMod
}
\description{
Constructor for xsl_model S3 class

xslMod S3 class

Constructor for xslMod S3 class
}
