% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSLmodels-package.R
\docType{package}
\name{XSLmodels-package}
\alias{XSLmodels}
\alias{XSLmodels-package}
\title{XSLmodels: Run and optimize cross-situational word learning models}
\description{
A collection of cognitive models of cross-situational word learning, together with methods and datasets for fitting and cross-validating them against experimental and corpus data, and for plotting the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kachergis.github.io/XSLmodels/}
  \item \url{https://github.com/kachergis/XSLmodels}
  \item Report bugs at \url{https://github.com/kachergis/XSLmodels/issues}
}

}
\author{
\strong{Maintainer}: George Kachergis \email{george.kachergis@gmail.com} (\href{https://orcid.org/0000-0003-4153-4167}{ORCID})

Authors:
\itemize{
  \item George Kachergis \email{george.kachergis@gmail.com} (\href{https://orcid.org/0000-0003-4153-4167}{ORCID})
  \item Mika Braginsky \email{mika.br@gmail.com} (\href{https://orcid.org/0000-0001-9039-3220}{ORCID})
  \item Kevin Liang \email{kevliang@sas.upenn.edu}
}

}
\keyword{internal}
