% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bayesian_decay.R
\name{bayesian_decay}
\alias{bayesian_decay}
\title{Bayesian decay model}
\usage{
bayesian_decay(alpha, delta, chdec)
}
\arguments{
\item{alpha}{Decay for word/object non-co-occurrence (0.1, 0.5, 0.9)}

\item{delta}{Multiplier for word/object co-occurrence (1 for no increase)}

\item{chdec}{Decision parameter (e.g., 1)}
}
\value{
An object of class xslMod
}
\description{
Bayesian decay model
}
\examples{
mod <- bayesian_decay(alpha = 0.5, delta = 1, chdec = 1)
xsl_run(mod, get_example_ambiguous_condition())
}
