% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fazly.R
\name{fazly}
\alias{fazly}
\title{Fazly et al. 2010 probablistic associative model}
\usage{
fazly(lambda, beta, alpha = 10, epsilon = 0.001, theta = NULL)
}
\arguments{
\item{lambda}{Small smoothing factor (e.g. 1e-5)}

\item{beta}{Upper bound on number of symbol types to expect (e.g. 8500)}

\item{alpha}{Fixed smoothing parameter (e.g. 10)}

\item{epsilon}{Fixed smoothing parameter (e.g. 0.001)}

\item{theta}{Threshold to determine when associations enter the lexicon
(defaults to NULL, i.e. model variant without thresholding).}
}
\value{
An object of class xslMod
}
\description{
Fazly et al. 2010 probablistic associative model
}
\examples{
mod <- fazly(lambda = 1e-5, beta = 8500)
xsl_run(mod, get_example_ambiguous_condition())

mod_t <- fazly(lambda = 1e-5, beta = 8500, theta = 0.7)
xsl_run(mod_t, get_example_ambiguous_condition())
}
