% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_example_unambiguous_condition}
\alias{get_example_unambiguous_condition}
\title{Get example unambiguous condition}
\usage{
get_example_unambiguous_condition()
}
\description{
This function creates an example condition with three trials and two word-object pairs
per trial. Each word-object pair appears twice, and so if the most frequently co-occurring
referent for each word is selected then all three words would be learned.
}
