% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_perf}
\alias{get_perf}
\title{Calculate Luce Choice (Proportion Correct) for Each Item in a Model Knowledge Matrix}
\usage{
get_perf(m)
}
\arguments{
\item{m}{A square matrix representing the model's knowledge, where rows
correspond to words and columns correspond to referents. The diagonal
elements represent correct associations, and off-diagonal elements represent
incorrect associations.}
}
\value{
A named numeric vector where each element corresponds to an
item in the matrix. The value of each element represents the proportion of
correct selections (Luce choice) for that item, calculated as the ratio of the
correct association (diagonal element) to the total associations for that item.
}
\description{
This function computes the Luce choice, or the proportion of correct selections,
for each item in a given model knowledge matrix. It assesses the probability of
correctly identifying each referent based on the knowledge matrix, providing a
measure of model performance per item.
}
