% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_roc}
\alias{get_roc}
\title{Calculate Receiver Operating Characteristic (ROC) Scores for a Model
Association Matrix}
\usage{
get_roc(mdat, fscores_only = T, plot = F, gold_lexicon = c())
}
\arguments{
\item{mdat}{A matrix representing the model association matrix with words as
rows and referents as columns.}

\item{fscores_only}{Logical; if TRUE, only the f-scores are returned for each
threshold. If FALSE, a dataframe including precision, recall, and f-scores
for each threshold is returned.}

\item{plot}{Logical; if TRUE, a plot of the ROC curve is generated and
displayed.}

\item{gold_lexicon}{Optional; a data frame or list where each row/element
represents a word-object pair in the gold lexicon. If provided, it is used
for more accurate calculation of true positives, false positives, and false
negatives.}
}
\value{
A dataframe or tibble with columns for the threshold, f-score,
precision, and recall. If \code{fscores_only} is TRUE, only f-scores are
returned for each threshold from 0 to 1 in increments of 0.01. If
\code{fscores_only} is FALSE, the dataframe includes precision and recall along
with f-scores for each threshold.
}
\description{
This function computes Receiver Operating Characteristic (ROC) scores for a
given model association matrix. It evaluates the performance of the model at
various thresholds, providing metrics like f-scores, precision, and recall.
The function can operate with a range of thresholds and optionally consider a
gold lexicon for calculating true and false positives and negatives. The
result is a comprehensive assessment of model performance over a continuum of
classification thresholds.
}
