% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_guess-and-test.R
\name{guess_and_test}
\alias{guess_and_test}
\title{Guess and test model}
\usage{
guess_and_test(f, sa)
}
\arguments{
\item{f}{forget at retrieval}

\item{sa}{prob of storage (slow learning down)}
}
\value{
An object of class xslMod
}
\description{
Guess and test model
}
