% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_kachergis.R
\name{kachergis}
\alias{kachergis}
\title{Kachergis 2012}
\usage{
kachergis(X, B, C, stochastic = FALSE)
}
\arguments{
\item{X}{Associative weight to distribute}

\item{B}{Weighting of uncertainty vs. familiarity}

\item{C}{Decay}

\item{stochastic}{}
}
\value{
An object of class xslMod
}
\description{
Kachergis 2012
}
