% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kachergis_initial_accuracy}
\alias{kachergis_initial_accuracy}
\title{Kachergis, Grimmick, & Gureckis initial accuracy experiment data}
\format{
An object of class \code{list} of length 2.
}
\usage{
kachergis_initial_accuracy
}
\description{
A list of two \code{xslData} objects ("High Initial Accuracy" and "Low Initial
Accuracy") from an unpublished MTurk cross-situational word learning
experiment manipulating how many of 18 word-object pairs are "switched"
(i.e. studied with the wrong partner) relative to an initial
familiarization phase. Not included in \link{xsl_datasets} (see
\code{data-raw/add_kachergis_initial_accuracy.R} for construction details and
\code{tests/bakeoff_comparison/kachergis_initial_accuracy_fit.R} for an example
model comparison).
}
\keyword{datasets}
