% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mafc_test}
\alias{mafc_test}
\title{Evaluate m-alternative forced choice test}
\usage{
mafc_test(mperf, test)
}
\arguments{
\item{mperf}{A matrix representing model memory with words as rows and
referents as columns.}

\item{test}{A list representing the test trials, each containing a word and
its associated referents.}
}
\value{
A vector with the probability of choosing the correct object, given
each word.
}
\description{
This function evaluates a given set of test trials using the provided model
memory matrix (word x referent). Each test trial is assumed to present one
word and a set of referents of size less than the width of the model memory
matrix.
}
