% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_propose-but-verify.R
\name{PBV}
\alias{PBV}
\title{Trueswell et al. 2013 propose-but-verify model}
\usage{
PBV(alpha, alpha_increase)
}
\arguments{
\item{alpha}{initial association strength}

\item{alpha_increase}{learning rate}
}
\value{
An object of class xslMod
}
\description{
Trueswell et al. 2013 propose-but-verify model
}
