% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Calculates Shannon entropy of a supplied vector, after normalizing it}
\usage{
shannon_entropy(p)
}
\arguments{
\item{p}{Numeric vector of probabilities}
}
\description{
Calculates Shannon entropy of a supplied vector, after normalizing it
}
