% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslMod-class.R
\name{update_params}
\alias{update_params}
\title{Update parameters of xslMod}
\usage{
update_params(xsl_mod, params)
}
\arguments{
\item{xsl_mod}{Object of class xslMod}

\item{params}{Named list of parameters}
}
\value{
An object of class xslMod
}
\description{
Update parameters of xslMod
}
