% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{xsl_run}
\alias{xsl_run}
\alias{xsl_run.xslMod}
\title{Fit XSL model}
\usage{
xsl_run(model, conds, ...)

\method{xsl_run}{xslMod}(model, conds, stochastic = FALSE, sse_only = FALSE, verbose = FALSE)
}
\arguments{
\item{sse_only}{Logical; if TRUE, only the SSE will be returned.}

\item{model_name}{An object of class xslMod.}

\item{conditions}{A list representing the conditions under which the model is
run. Each condition should include training data (\code{train}) and human item
accuracy (\code{HumanItemAcc}).}

\item{print_perf}{Logical; if TRUE, model performance and SSE will be
printed.}
}
\value{
TODO
}
\description{
Fit XSL model
}
