% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{xsl_run}
\alias{xsl_run}
\title{Run XSL model}
\usage{
xsl_run(model, data, control = xsl_control())
}
\arguments{
\item{model}{An object of class xslMod.}

\item{data}{An object (or list of objects) of class xslData.}

\item{control}{Control arguments returned by \code{xsl_control()}.}
}
\value{
List of fits, sse, unweighted_sse
}
\description{
Run XSL model
}
