% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{xsl_run}
\alias{xsl_run}
\alias{xsl_run.xslMod}
\title{Fit XSL model}
\usage{
xsl_run(model, data, ...)

\method{xsl_run}{xslMod}(model, data, sse_only = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{An object of class xslMod.}

\item{data}{An object of class xslData.}

\item{sse_only}{Logical; if TRUE, only the SSE will be returned.}

\item{print_perf}{Logical; if TRUE, model performance and SSE will be
printed.}
}
\value{
TODO
}
\description{
Fit XSL model
}
