% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{xsl_run}
\alias{xsl_run}
\title{Run XSL model}
\usage{
xsl_run(model, data, reps = 1, verbose = FALSE)
}
\arguments{
\item{model}{An object of class xslMod.}

\item{data}{An object of class xslData.}

\item{reps}{Number of times to repeat training (defaults to 1).}

\item{sse_only}{Logical; if TRUE, only the SSE will be returned.}

\item{print_perf}{Logical; if TRUE, model performance and SSE will be
printed.}
}
\value{
TODO
}
\description{
Run XSL model
}
