% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslControl-class.R
\name{xslControl-class}
\alias{xslControl-class}
\alias{xslControl}
\alias{new_xslControl}
\title{xslControl S3 class}
\usage{
xslControl(
  reps = 1,
  start_matrix = NULL,
  test_noise = 0,
  n_sim = 500,
  verbose = FALSE
)

new_xslControl(x = list())
}
\arguments{
\item{reps}{Number of times to repeat training}

\item{start_matrix}{Starting matrix}

\item{test_noise}{Test noise}

\item{n_sim}{Number of simulations for stochastic models}

\item{verbose}{Verbosity}

\item{x}{List with elements train, test, accuracy, n_subj, label, condition}
}
\value{
An object of class xslControl
}
\description{
Control arguments for \code{xsl_run()}
}
