% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslData-class.R
\name{xslData-class}
\alias{xslData-class}
\alias{xslData}
\title{xslData S3 class}
\usage{
xslData(
  train = list(),
  test = list(),
  accuracy = numeric(),
  n_subj = numeric(),
  label = character(),
  condition = character()
)
}
\arguments{
\item{label}{}
}
\value{
An object of class xslMod
}
\description{
xslData S3 class
}
