% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslFit-class.R
\name{xslFit-class}
\alias{xslFit-class}
\alias{xslFit}
\title{xslFit S3 class}
\usage{
xslFit(perf = matrix(), matrix = matrix(), traj = list(), sse = double())
}
\arguments{
\item{perf}{Perf}

\item{matrix}{Matrix}

\item{traj}{Traj}

\item{sse}{SSE}
}
\value{
An object of class xslFit
}
\description{
xslFit S3 class
}
