% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslData-class.R, R/xslFit-class.R,
%   R/xslMod-class.R
\name{new_xslData}
\alias{new_xslData}
\alias{new_xslFit}
\alias{xslMod-class}
\alias{xslMod}
\alias{new_xslMod}
\title{Constructor for xsl_model S3 class}
\usage{
new_xslData(x = list())

new_xslFit(x = list())

xslMod(
  name = character(),
  description = character(),
  model,
  params = numeric(),
  stochastic = logical()
)

new_xslMod(x = list())
}
\arguments{
\item{name}{Name}

\item{description}{Description}

\item{model}{Model fitting function}

\item{params}{List of parameters}
}
\value{
An object of class xslMod
}
\description{
Constructor for xsl_model S3 class

Constructor for xsl_model S3 class

xslMod S3 class

Constructor for xsl_model S3 class
}
