% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSLmodels-package.R
\docType{package}
\name{XSLmodels-package}
\alias{XSLmodels}
\alias{XSLmodels-package}
\title{XSLmodels: Run and optimize cross-situational word learning models}
\description{
This package has a collection of cognitive models of cross-situational word learning models, as well as methods and datasets for optimizing them, and plotting the results.
}
\author{
\strong{Maintainer}: George Kachergis \email{george.kachergis@gmail.com} (\href{https://orcid.org/0000-0003-4153-4167}{ORCID})

Authors:
\itemize{
  \item Kevin Liang \email{kevliang@sas.upenn.edu} (\href{https://orcid.org/ORCID-NUMBER}{ORCID})
}

}
\keyword{internal}
